Make some fixes and add unit tests for decoding

+++ app/models/MicrophoneDataModel.scala
@@ -2,9 +2,9 @@ package models
 
 case class MicrophoneDataModel(id: Int,
                                isFix: Boolean,
-                               volume: Short,
-                               low: Short,
-                               med: Short,
-                               high: Short) extends DataModel{
+                               volume: Int,
+                               low: Int,
+                               med: Int,
+                               high: Int) extends DataModel{
   def getId = id
 }
